feat: route to nodes even if slightly behind max height (#68)
We want to be able to tolerate some lag in certain situations. For example, we might want to favour our own node over a node provider even if we're slightly behind head. This PR keeps routing to nodes as long as they are within 10 blocks of the max height.

# Test Plan

Added unit tests for new filter.