fix: turn JSON RPC request body ID to pointer
JSON RPC requests can omit the ID field to signal that the request is a notification: https://www.jsonrpc.org/specification#notification.
These notification requests are not expected to have a response.

The SingleRequestBody model in the jsonrpc module uses the `omitempty`
tag on the ID field.  The default "zero" value for an int64 in Golang is
zero.  This means that any JSONRPC requests with ID of zero will not
have the ID field present.  This causes the request to look like a
notification, and the resulting response will have content-length of
zero as well.

The graph-node uses a library called rust-web3 to make the JSONRPC
requests to node providers.  This library ends up making requests that
have ID set to 0 in the request.  When pointed to the node-gateway, the
request hits the bug explained above and the response body is empty.
The graph-node is not able to deserialize the empty response body
correctly and we see the error: `failed to deserialize response: data
did not match any variant of untagged enum Output`.  Interally,
graph-node uses a model named `Output` to deserialize the JSONRPC
response, hence why `untagged enum Output` shows up in the error.

To prevent this, we can use a pointer for the ID field in the request
model.  This prevents omitempty from removing the ID field when it is
zero.  This approach is more favorable because we do not have control
over the JSONRPC ID used in graph-node.

Fixed the tests to reflect the new field type.  Changed the
SingleResponseBody ID to also be int64 to be more consistent.  A valid
JSONRPC response will never have ID not set, so it is OK to not have it
as a pointer.