fix: support null result in JSONRPC response (#107)
* fix: support null result in JSONRPC response

We have seen several cases where the node-gateway responds to a
graph-node with a missing `result` field. This results in an error
message: `"failed to deserialize response: data did not match any
variant of untagged enum Output"` that is difficult to debug. It also
shows up in logs on customer subgraph dashboards, which leads to
confusion for the end-user.

The main issue is that nodes will respond with `result` as null, which
will get removed during unmarshaling due to the `omitempty` tag. In our
case, we want to differentiate between an implicit null and an explicit
null. We don't want the `result` field when it is not set(implicit),
but we DO want `"result": null`(explicit). This setup is not easy to do
out of the box with JSON tags. We also can't use the pointer trick
because nil is still the zero value for pointers.

Couple approaches we can take here. First is customizing
unmarshaling logic, which can get complex to maintain over time. Second
approach is using the `json.RawMessage` type, which basically tells the
marshal/unmarshal logic that we want the raw value. An explicit `null`
value still gets past the unmarshal while an unset(implicit) value is
handled properly by the omitempty tag.

* fix: json.RawMessage in unit tests