Use json.Number to support both ints & strings IDs
ID numbers can be either an int or a string when being sent from
clients. Node-gateway previously only accepted ints, but this can cause
problems when using clients such as Metamask.

This change alters the type to be a number, which allows either to be
accepted from clients. As a consequence, the gateway will always return
only ints in response messages even if the client originally sent a
string ID.