fix: error handling with request coalescing
The golang-redis/cache library supports request coalescing, which means
that multiple identical requests to the cache will only result in a
single request to origin.

However, this means that errors can also be returned in the scenario.
After looking at logs, this is very common for node-gateway because the
graph-nodes are constantly getting null results from
eth_getTransactionReceipt calls for new transactions.

Certain errors that come from request coalescing can be handled(null
result and JSON RPC Errors). We have to fake a response and modify the
response body to match the original request.