feat: add cached tag for request metrics (#120)
Cleaned up some of the JSONRPC request metrics that are not used by
dashboards right now. They have most of the same information as the
corresponding upstream request metric. The only one that I could not
remove as easily was the JSONRPC request error metric, which is more
specific to the Error field in the JSON RPC response. This has a real
use compared to the general upstream request error metric that measures
non-200s.

The cached variable that is passed up to the metric does not exactly
mean that it was a real cache hit to redis. The go-redis/cache library
does not expose its internal cached state, so we can only try our best.
If the Do() function is entered at all, we assume that the result was
not cached. Otherwise, we assume that the result was cached.
Technically, this means that even coalesced requests will be classified
as cached as well. I think this is OK, since it means that there was no
extra request to the origin.